Add react component for currency input
The component is used by pochtron to credit account. I change from
react-input-mask such that the behavior is the same as the price input
in alcohol edit. The CurrencyInput component is a traduction in React of
this code in Typescript https://github.com/larkintuckerllc/react-currency-input/blob/master/src/App.tsx